`@remotion/studio`: Split inspector panel components
